backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── MongoDB async ─────────────────────────────────────────────────────────────
motor==3.3.2
pymongo==4.6.1

# ── HTTP client (appels LLM, tous providers) ──────────────────────────────────
httpx==0.27.2

# ── PDF extraction ────────────────────────────────────────────────────────────
pymupdf==1.24.10

# ── Auth / Sécurité ───────────────────────────────────────────────────────────
PyJWT==2.9.0
python-multipart==0.0.12

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.9.2

pymupdf 
pytesseract 
pillow
